% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_fastpot}
\alias{fp_fastpot}
\title{Fast Potential Calulation}
\usage{
fp_fastpot(
  knownpts,
  unknownpts,
  var = "v",
  fun = "e",
  span = 2000,
  beta = 2,
  limit = 10000,
  ncl = 3,
  size = 500
)
}
\arguments{
\item{knownpts}{sf point, known pts}

\item{unknownpts}{sf point, known points}

\item{var}{variable name}

\item{fun}{friction function, "e" (expon) or "p" (pareto)}

\item{span}{span}

\item{beta}{beta}

\item{limit}{limit}

\item{ncl}{number of core}

\item{size}{size chunk}
}
\value{
a vector
}
\description{
stewart with cuttoff (limit) and parrallel
}
\examples{
\dontrun{
library(sf)
library(cartography)
library(SpatialPosition)

mtq <- st_read(system.file("gpkg/mtq.gpkg", package="cartography"))
grid <- st_make_grid(mtq, cellsize = 500)
grid <- st_sf(id = 1:length(grid), geom = grid)
set.seed(1)
size = 50000
pt <- st_sample(grid, size = size)
pt <- st_sf(id = 1:length(pt), geom = pt)
cc <- st_coordinates(pt)
pt$v <- runif(n = size, 10000, 20000)
pt$v <- pt$v + cc[,1] + cc[,2] / 1000

system.time(
  pot2 <- mcStewart(pt, grid, varname = "v", span = 2000, beta = 2,
                    returnclass = "sf", cl = 3, size  = 100)
)
system.time(
  pot <- fp_fastpot(pt, grid, var = c('v'), fun="e",beta = 2, limit = 10000,
                    ncl = 3, size = 100)
)

plot(pot,pot2$OUTPUT, asp = 1)
lm(pot ~pot2$OUTPUT)$coeff

grid$pot <- pot
par(mfrow = c(1,2))
choroLayer(pot2, var = "OUTPUT", border = NA)
choroLayer(grid, var = "pot", border = NA)
}
}
